Removed some overkill comments

+++ src/Core/Config.cpp
@@ -20,10 +20,10 @@ bool Config::Save()
     m_Config["controls"]["showcontrols"] = Context::DisplayControlsTexts ? "true" : "false";
     m_Config["position"]["x"] = std::to_string(static_cast<long double>(Layout::X));
     m_Config["position"]["y"] = std::to_string(static_cast<long double>(Layout::Y));
-    m_Config["color"]["r"] = std::to_string(static_cast<long long>(D3DCOLOR_GETRED(Layout::Color)));
-    m_Config["color"]["g"] = std::to_string(static_cast<long long>(D3DCOLOR_GETGREEN(Layout::Color)));
-    m_Config["color"]["b"] = std::to_string(static_cast<long long>(D3DCOLOR_GETBLUE(Layout::Color)));
-    m_Config["color"]["a"] = std::to_string(static_cast<long long>(D3DCOLOR_GETALPHA(Layout::Color)));
+    m_Config["color"]["r"] = std::to_string(static_cast<uint64_t>(D3DCOLOR_GETRED(Layout::Color)));
+    m_Config["color"]["g"] = std::to_string(static_cast<uint64_t>(D3DCOLOR_GETGREEN(Layout::Color)));
+    m_Config["color"]["b"] = std::to_string(static_cast<uint64_t>(D3DCOLOR_GETBLUE(Layout::Color)));
+    m_Config["color"]["a"] = std::to_string(static_cast<uint64_t>(D3DCOLOR_GETALPHA(Layout::Color)));
 
     // Generate the config file
     return m_ConfigFile.generate(m_Config);
